requirements.txt
numpy==1.13.3
Pillow==5.0.0
